requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.6
cmake==3.26.4
coloredlogs==15.0.1
commonmark==0.9.1
cryptography==41.0.1
dataclasses-json==0.5.9
Deprecated==1.2.14
duckdb==0.8.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.99.1
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
graphlib-backport==1.0.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.264
langchainplus-sdk==0.0.19
langsmith==0.0.22
lit==16.0.6
llama-index==0.8.1.post1
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandoc==2.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
plumbum==1.8.2
ply==3.11
posthog==3.0.1
protobuf==4.23.4
pulsar-client==3.2.0
pyarrow==12.0.1
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.10
pydeck==0.8.0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
pypdf==3.12.0
pypdf2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.30.2
regex==2023.6.3
requests==2.29.0
rfc3986==1.5.0
rich==13.0.1
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typer==0.7.0
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.7.12
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
validators==0.21.2
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
youtube-transcript-api==0.6.1
yt-dlp==2023.6.22
zipp==3.15.0
zstandard==0.21.0
